Switch CHANGELOG from asciidoc to markdown
This was an experiment in other markup languages. Markdown has won,
however, and keeping this single document now that most other
documentation is in markdown just doesn't make sense.

Also found that the original Keep A Changelog template had links to
each of the release versions in the version section headers, so
added those in.