Add EventCache for win event HANDLEs.
EventCache will hold the HANDLE for a read/write when not in use or will
produce a new handle for each thread when multiple threads try to
read/write at the same time. The cache holds a single value and will
deallocate extra handles when there is already a handle stored. We have
one cache for a read_handle and one cache for a write_handle.

The expectation is that in the normal case, at most one thread will be
reading/writing at a time, and so at most one handle will be created for
each of read and write. But in abnormal cases, when multiple threads try
to read/write at the same time, we auto-create and auto-close extra
handles as needed.

Fix the into_raw_handle/into_raw_fd to drop other fields that were
previously leaked.