Upgrade clap to 4.0
Remove the patch version in the dependency definition to get minor fixes
automatically. This shouldn't be a problem as clap should be API stable
across patch versions.

There were some deprecations in the 3.2.0 and 4.0.0 release which were
also fixed in this patch:

	ArgMatches::value_of   => ArgMatches::get_one::<T>
	ArgMatches::is_present => ArgMatches::get_flag
	Arg::takes_value(true) => Arg::num_args(1)
	Arg::validator         => Arg::value_parser
	Arg::possible_values   => Arg::value_parser(PossibleValuesParser::new())

For more information please see the corresponding changelog at
	https://github.com/clap-rs/clap/blob/master/CHANGELOG.md

Signed-off-by: Richard Leitner <richard.leitner@skidata.com>