Upgrade bitflags to version from 1.3.1 to 2.0.y
This allows 2.0.y bitflags versions. There are no obvious changes
necessary to our code base according to the release notes [1]. Therefore
this should be fine.

The range is given to avoid having 2 bitflags versions as clap
still depends on the 1.3.2 version. To resolve this issue there is an
extra PR open [2]. As soon as this got merged we can drop the 1.3.y
support.

We are excluding 2.1.y releases as bitflags may change the MSRV in minor
releases.

[1] https://github.com/bitflags/bitflags/releases
[2] https://github.com/serialport/serialport-rs/pull/76

Signed-off-by: Richard Leitner <dev@g0hl1n.net>