Increase MSRV coverage for supported platforms in CI
Let's be nice and check the MSRV (at least once) for each supported
platform according to `README.md`. We differentiate Linux further with
respect to udev/musl, so it gets a special treatment.

Our examples currently do not build with MSRV. Let's work around this by
not building them for the new MSRV check msrv-x86_64-unknown-linux-musl
for now.