Fix nightly (#169)
* Fix nightly lint causing compilation to fail

* Relax checks on unused imports

This is definitely worth a warning but should not make builds fail for
users. We deny warnings in CI and sort this out before integrating.

See issue #169 for more details.

* Don't let nightly builds fail CI

* Clean up superfluous explicit import

Current nightly considers use nix::self as an unused import as nix is
already through the extern prelude (since edition 2018) and generates a
warning. Remove the explicit import to silence it.

See:

* Output of 'cargo --verbose check'
* https://doc.rust-lang.org/stable/reference/names/preludes.html#extern-prelude

---------

Co-authored-by: Ross Smyth <rsmyth@electrocraft.com>
Co-authored-by: Christian Meusel <christian.meusel@posteo.de>