Always look up UsbPortInfo on macOS from IOUSBHostInterface
USB interface number information is currently missing macOS (with
feature usbportinfo-interface). The required information is not provided
in IOUSBHostDevice but along with all other information of interest in
IOUSBHostInterface.

As it is planned to remove the feature gate for  UsbPortInfo::interface
with the next major release, IOUSBHostInterface is the dict to use for
looking up information for USB serial ports. There is no point in using
different lookups in the meanwhile.