Don't ingnore errors when setting baud rate
"Don't you think we should tell them that this baud rate is not
supported? Nah. Just smile and wave boys smile, smile and wave." While
this makes great movies, this does not make setting a knowingly
unsupported baud rate a great experience. At least when in comes to
debugging.

Play nice and report this and other errors related to baud rates back.