Remove i686-pc-windows-gnu from CI builds
Builds for this target are broken on the GitHub Windows 2025 runner,
this target got recently demoted from tier 1 to 2 with Rust 1.88, I
can't reproduce this issue on my windows machine, and looking into the
actual build in CI is cumbersome. So I will remove this target from CI
builds in favor of having a meaningful overall build status.

If someone feels like doing the deep dive into this issue, I'm happy to
lend a hand but I don't feel like leading this endeavor.