Revert automatically setting DTR when opening a port
This behavior looked promising in the first place to get some Arduino
devices to send data at all (like in #29 and #204). But after shipping
this as the new default, several issues were reported from other Arduino
devices and other places (see issue #243 and #251).

So I'm finally pulling the plug on this and revert setting DTR on open
by default. I'm sorry for the back and forth and all the inconvenience
this caused!