Harmonize error kind returned when port locking fails
Previously, when we only supported TIOCEXCL, the error conversion from
nix::Error to serialport::Error did not care for EBUSY and converted this to
ErrorKind::Unknown with just the message hinting at the port being busy (due to
already being locked).

This was in contrast to our documentation of our POSIX error kinds states that
ErrorKind::NoDevice will be returned in case that port locking does not
succeed.

The recently introduced convenience wrappers for flock apparently read this
documentation and return ErrorKind::NoDevice in this case. Well, let's return
the documented error kind for failed TIOCEXCL locking too.